AI/pdf_search/requirements.txt
PyPDF2==3.0.1
openai==1.12.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0 